generate and add additional targets after processing all standard targets
We generate extra link targets both for backwards compatibility, as well
as for the first word to make linking to functions easier in many cases.
Previously, these targets were added as the document was being
processed. But the first word of a heading can easily conflict with a
later heading that should take priority. This means the extra targets
need to be added only after the entire document has been processed.

Add a new role that buffers the entire output and allows filtering it.
Also provide a method that is called whenever new output added. This
allows the code adding the extra targets to collect all of the
information needed to add the targets as the document is processed, and
then when finished inject the new targets.